Changed: Minors server changes from Hg

+++ src/entities_game_service/phrase_manager/special_power_phrase.cpp
@@ -708,7 +708,7 @@ bool CSpecialPowerPhrase::validate()
 	_Targets.push_back(entity->getTargetDataSetRow());
 
 	// validate all powers
-	const uint nbPowers = _Powers.size();
+	const uint nbPowers = (uint)_Powers.size();
 	for (uint i = 0 ; i < nbPowers ; ++i)
 	{
 #ifdef NL_DEBUG
@@ -776,7 +776,7 @@ void CSpecialPowerPhrase::apply()
 void CSpecialPowerPhrase::applyPowers()
 {
 	// apply powers
-	const uint nbPowers = _Powers.size();
+	const uint nbPowers = (uint)_Powers.size();
 	for (uint i = 0 ; i < nbPowers ; ++i)
 	{
 	#ifdef NL_DEBUG
